Implementação dos botões de filtro refatorada, para reduzir a complexidade do código conforme apontado pelo lint
