+++ src/components/RecipeDoneCard.jsx
@@ -1,201 +1,16 @@
 import React, { useContext } from 'react';
-import { Link } from 'react-router-dom';
 import Context from '../context/Context';
-import shareIcon from '../images/shareIcon.svg';
+import RecipeDoneCardFilter from './RecipeDoneCardFilter';
+import RecipeDoneCardAll from './RecipeDoneCardAll';
 
 export default function RecipeDoneCard() {
   const { allRecipesDone, filterRecipeDone } = useContext(Context);
 
-  function renderTags(stringTags) {
-    if (stringTags) {
-      const separator = /,\s*/; // ,' '
-      const allTags = stringTags.split(separator); // cria um array de tags
-      return allTags;
-    }
-    return '';
-  }
+  // RETORNA COMPONENTE COM AS RECEITAS FILTRADAS
+  if (filterRecipeDone.length !== 0) return (<RecipeDoneCardFilter />);
 
-  // CÓDIGO DO FILTRO
-  if (filterRecipeDone.length !== 0) {
-    return (
-      filterRecipeDone
-        .map((recipe, index) => (
-          <div
-            key={ index }
-            data-testid={ `${index}-recipe-card` }
-            className="recipe-done-card"
-          >
-            <div className="recipe-card-img">
-              <Link
-                to={
-                  recipe.type === 'Meal'
-                    ? `/comidas/${recipe.idMeal}`
-                    : `/bebidas/${recipe.idDrink}`
-                }
-              >
-                <img
-                  data-testid={ `${index}-horizontal-image` }
-                  src={ recipe[`str${recipe.type}Thumb`] }
-                  alt="thumbnail"
-                />
-              </Link>
-            </div>
-
-            <div className="recipe-card-category">
-              <div
-                data-testid={ `${index}-horizontal-top-text` }
-              >
-                {
-                  recipe.type === 'Meal'
-                    ? `${recipe.strArea} - ${recipe.strCategory}`
-                    : `${recipe.strAlcoholic}`
-                }
-              </div>
-            </div>
-
-            <div className="recipe-card-title">
-              <Link
-                to={
-                  recipe.type === 'Meal'
-                    ? `/comidas/${recipe.idMeal}`
-                    : `/bebidas/${recipe.idDrink}`
-                }
-              >
-                <h4 data-testid={ `${index}-horizontal-name` }>
-                  { recipe[`str${recipe.type}`] }
-                </h4>
-              </Link>
-            </div>
-
-            <div className="recipe-card-date">
-              <div data-testid={ `${index}-horizontal-done-date` }>
-                { `Feita em: ${recipe.date}` }
-              </div>
-            </div>
-
-            <div className="recipe-card-tags">
-              <div
-                data-testid={ `${index}-${renderTags(recipe.strTags)[0]}-horizontal-tag` }
-              >
-                { renderTags(recipe.strTags)[0] }
-              </div>
-              <div
-                data-testid={ `${index}-${renderTags(recipe.strTags)[1]}-horizontal-tag` }
-              >
-                { renderTags(recipe.strTags)[1] }
-              </div>
-            </div>
-
-            <div className="recipe-card-share-btn">
-              <a
-                href={
-                  recipe.type === 'Meal'
-                    ? `/comidas/${recipe.idMeal}`
-                    : `/bebidas/${recipe.idDrink}`
-                }
-                src={ shareIcon }
-                data-testid={ `${index}-horizontal-share-btn` }
-                onClick={ () => {} }
-              >
-                <img src={ shareIcon } alt="share" />
-              </a>
-            </div>
-
-          </div>
-        )) // end map()
-    ); // end return()
-  } // end if
-
-  // CÓDIGO QUE MOSTRA TODAS AS RECEITAS FEITAS
-  if (allRecipesDone.length !== 0) {
-    return (
-      allRecipesDone
-        .map((recipe, index) => (
-          <div
-            key={ index }
-            data-testid={ `${index}-recipe-card` }
-            className="recipe-done-card"
-          >
-            <div className="recipe-card-img">
-              <Link
-                to={
-                  recipe.type === 'Meal'
-                    ? `/comidas/${recipe.idMeal}`
-                    : `/bebidas/${recipe.idDrink}`
-                }
-              >
-                <img
-                  data-testid={ `${index}-horizontal-image` }
-                  src={ recipe[`str${recipe.type}Thumb`] }
-                  alt="thumbnail"
-                />
-              </Link>
-            </div>
-
-            <div className="recipe-card-category">
-              <div
-                data-testid={ `${index}-horizontal-top-text` }
-              >
-                {
-                  recipe.type === 'Meal'
-                    ? `${recipe.strArea} - ${recipe.strCategory}`
-                    : `${recipe.strAlcoholic}`
-                }
-              </div>
-            </div>
-
-            <div className="recipe-card-title">
-              <Link
-                to={
-                  recipe.type === 'Meal'
-                    ? `/comidas/${recipe.idMeal}`
-                    : `/bebidas/${recipe.idDrink}`
-                }
-              >
-                <h4 data-testid={ `${index}-horizontal-name` }>
-                  { recipe[`str${recipe.type}`] }
-                </h4>
-              </Link>
-            </div>
-
-            <div className="recipe-card-date">
-              <div data-testid={ `${index}-horizontal-done-date` }>
-                { `Feita em: ${recipe.date}` }
-              </div>
-            </div>
-
-            <div className="recipe-card-tags">
-              <div
-                data-testid={ `${index}-${renderTags(recipe.strTags)[0]}-horizontal-tag` }
-              >
-                { renderTags(recipe.strTags)[0] }
-              </div>
-              <div
-                data-testid={ `${index}-${renderTags(recipe.strTags)[1]}-horizontal-tag` }
-              >
-                { renderTags(recipe.strTags)[1] }
-              </div>
-            </div>
-
-            <div className="recipe-card-share-btn">
-              <a
-                href={
-                  recipe.type === 'Meal'
-                    ? `/comidas/${recipe.idMeal}`
-                    : `/bebidas/${recipe.idDrink}`
-                }
-                src={ shareIcon }
-                data-testid={ `${index}-horizontal-share-btn` }
-                onClick={ () => {} }
-              >
-                <img src={ shareIcon } alt="share" />
-              </a>
-            </div>
-
-          </div>
-        )) // end map()
-    ); // end return()
-  } // end if
+  // RETORNA COMPONENTE COM TODAS AS RECEITAS FEITAS
+  if (allRecipesDone.length !== 0) return (<RecipeDoneCardAll />);
 
   return null;
 } // end function
